feat(new-version): :sparkles: add `reset` file as a `mixin` to be used easily in `layer` CSS or not
Add `reset` into a `mixin` to be easily used into CSS layers or render it without using it. We also changed the version of the library as a break changes and make it as `v2.0.0`. We also removed unneeded styles for `noscript` tag. We removed also the text in `_config.scss` file of the `$no-script-text` variable as it will not be needed.